cache other values #17
