+++ src/tracktime.tsx
@@ -11,8 +11,9 @@ import {
   showToast,
 } from "@raycast/api";
 import { useEffect, useState } from "react";
-import { addTime, getEmployeeInfo, getPersonioToken } from "./api";
+import { addTime, getPersonioToken } from "./api/api";
 import moment from "moment-timezone";
+import { getEmployeeInfo } from "./api/employeeinfo";
 
 export default function TrackTime() {
   const [token, setToken] = useState("");
